chore(docs): remove docs/ folder
I would suggest that this could be removed. If people re-render locally, they will end up with a bunch of changes to files in docs/, which can make it harder to see what changes are made in a pull request, for example, as it is cluttered with HTML changes. Also, as the website is built using GitHub actions, you don't need to be committing update versions of docs/, and can end up with out-of-sync docs/ and published site. I tend not to commit docs/, and find it makes life easier, no issues! :)