fix(yaml/callout): fix rating callout + author YAML
Callout was not rendering as set to callout-info (not real) - changed to callout-note. Also, in YAML the "author" was being set to the person who wrote the page for the atlas, but I would suggest this feels misleading. When I read the pages, I thought it meant the author of the package/tool, especially given location of author details on page. I have suggested this is changed as such. But if you did still want to credit whoever wrote each page, I would maybe suggest that is inserted at the end of the page, .e.g Page authored by: Last updated:, or similar.