feat(status_reason): move status reason and status into title block (#56)
To avoid users being inconsistent with how they structure and where they place the rationale for their status in package, project or tool entries, this has been moved into the yml header under the pub-info section, using the key 'rationale-for-status'. This is then picked up by the title-block and displayed within the abstract box.